Fixed betty code style and variable type

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -7,12 +7,13 @@
  */
 size_t dlistint_len(const dlistint_t *h)
 {
-	int elem = 0;
-	while(h != NULL)
+	size_t elem = 0;
+
+	while (h != NULL)
 	{
 		if (h->n)
 			elem++;
 		h = h->next;
 	}
-	return(elem);
+	return (elem);
 }
